refator code schema
