+++ client/api_client.go
@@ -1,4 +1,4 @@
-package api
+package client
 
 import (
 	"bytes"
@@ -14,51 +14,65 @@ import (
 	"strings"
 	"time"
 
-	sdkmath "cosmossdk.io/math"
 	hashlib "github.com/bnb-chain/greenfield-common/go/hash"
 	httplib "github.com/bnb-chain/greenfield-common/go/http"
-	chainclient "github.com/bnb-chain/greenfield/sdk/client"
+	sdkerror "github.com/bnb-chain/greenfield-go-sdk/pkg/error"
+	"github.com/bnb-chain/greenfield-go-sdk/pkg/utils"
+	types3 "github.com/bnb-chain/greenfield-go-sdk/types"
 	sdkclient "github.com/bnb-chain/greenfield/sdk/client"
-	"github.com/bnb-chain/greenfield/types/common"
+	"github.com/bnb-chain/greenfield/sdk/keys"
+	types2 "github.com/bnb-chain/greenfield/sdk/types"
 	permTypes "github.com/bnb-chain/greenfield/x/permission/types"
 	storageTypes "github.com/bnb-chain/greenfield/x/storage/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/rs/zerolog/log"
-
-	"github.com/bnb-chain/greenfield-go-sdk/client"
-	account2 "github.com/bnb-chain/greenfield-go-sdk/pkg/account"
-	sdkerror "github.com/bnb-chain/greenfield-go-sdk/pkg/error"
-	"github.com/bnb-chain/greenfield-go-sdk/pkg/types"
-	"github.com/bnb-chain/greenfield-go-sdk/pkg/utils"
+	"google.golang.org/grpc"
 )
 
-type Client struct {
+type Client interface {
+	Basic
+	Bucket
+	Object
+	Group
+	Challenge
+	Account
+}
+
+type client struct {
 	// chainClients
 	chainClient *sdkclient.GreenfieldClient
 	// tendermintClient
 	tendermintClient *sdkclient.TendermintClient
 	// httpClient
 	httpClient *http.Client
-	// account
-	account *account2.Account
 	// spEndpoints
 	spEndpoints map[string]*url.URL
 
+	// TODO(leo): Unused variables
 	userAgent string
-	host      string
-	Secure    bool
+
+	host   string
+	Secure bool
+}
+
+type Option struct {
+	// keyManager is the manager used for generating and managing keys.
+	KeyManager keys.KeyManager
+	// grpcDialOption is the list of grpc dial options.
+	GrpcDialOption grpc.DialOption
+	Secure         bool
 }
 
 // New - instantiate greenfield chain with options
-func New(chainID string, grpcAddress, rpcAddress string, gnfdopts ...chainclient.GreenfieldClientOption) (*Client, error) {
+func New(chainID string, grpcAddress, rpcAddress string, option *Option) (Client, error) {
 	tc := sdkclient.NewTendermintClient(rpcAddress)
-	cc := sdkclient.NewGreenfieldClient(grpcAddress, chainID, gnfdopts...)
+	cc := sdkclient.NewGreenfieldClient(grpcAddress, chainID, sdkclient.WithKeyManager(option.KeyManager), sdkclient.WithGrpcDialOption(option.GrpcDialOption))
 
-	c := &Client{
+	c := client{
 		chainClient:      cc,
 		tendermintClient: &tc,
 		httpClient:       &http.Client{},
-		userAgent:        types.UserAgent,
+		userAgent:        types3.UserAgent,
 	}
 	// fetch sp endpoints info from chain
 	spInfo, err := c.GetSPAddrInfo()
@@ -67,11 +81,11 @@ func New(chainID string, grpcAddress, rpcAddress string, gnfdopts ...chainclient
 	}
 
 	c.spEndpoints = spInfo
-	return c, nil
+	return &c, nil
 }
 
 // getSPUrlByBucket route url of the sp from bucket name
-func (c *Client) getSPUrlByBucket(bucketName string) (*url.URL, error) {
+func (c *client) getSPUrlByBucket(bucketName string) (*url.URL, error) {
 	ctx := context.Background()
 	bucketInfo, err := c.HeadBucket(ctx, bucketName)
 	if err != nil {
@@ -97,7 +111,7 @@ func (c *Client) getSPUrlByBucket(bucketName string) (*url.URL, error) {
 }
 
 // getSPUrlByAddr route url of the sp from sp address
-func (c *Client) getSPUrlByAddr(address string) (*url.URL, error) {
+func (c *client) getSPUrlByAddr(address string) (*url.URL, error) {
 	if _, ok := c.spEndpoints[address]; ok {
 		return c.spEndpoints[address], nil
 	}
@@ -115,11 +129,38 @@ func (c *Client) getSPUrlByAddr(address string) (*url.URL, error) {
 	}
 }
 
+// requestMeta - contains the metadata to construct the http request.
+type requestMeta struct {
+	bucketName       string
+	objectName       string
+	urlRelPath       string     // relative path of url
+	urlValues        url.Values // url values to be added into url
+	Range            string
+	ApproveAction    string
+	TxnMsg           string
+	SignType         string
+	contentType      string
+	contentLength    int64
+	contentMD5Base64 string // base64 encoded md5sum
+	contentSHA256    string // hex encoded sha256sum
+	challengeInfo    types3.ChallengeInfo
+	userInfo         UserInfo
+}
+
+// SendOptions -  options to use to send the http message
+type sendOptions struct {
+	method           string      // request method
+	body             interface{} // request body
+	disableCloseBody bool        // indicate whether to disable automatic calls to resp.Body.Close()
+	txnHash          string      // the transaction hash info
+	isAdminApi       bool        // indicate if it is an admin api request
+}
+
 // newRequest constructs the http request, set url, body and headers
-func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta,
-	body interface{}, txnHash string, isAdminAPi bool, endpoint *url.URL, authInfo AuthInfo) (req *http.Request, err error) {
+func (c *client) newRequest(ctx context.Context, method string, meta requestMeta,
+	body interface{}, txnHash string, isAdminAPi bool, endpoint *url.URL, authInfo types3.AuthInfo) (req *http.Request, err error) {
 	// construct the target url
-	desURL, err := c.GenerateURL(meta.bucketName, meta.objectName, meta.urlRelPath, meta.urlValues, isAdminAPi, endpoint)
+	desURL, err := c.generateURL(meta.bucketName, meta.objectName, meta.urlRelPath, meta.urlValues, isAdminAPi, endpoint)
 	if err != nil {
 		return nil, err
 	}
@@ -132,7 +173,7 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 		if ObjectReader, ok := body.(io.Reader); ok {
 			reader = ObjectReader
 			if meta.contentType == "" {
-				contentType = types.ContentDefault
+				contentType = types3.ContentDefault
 			}
 		} else {
 			// the body content is xml type
@@ -140,7 +181,7 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 			if err != nil {
 				return nil, err
 			}
-			contentType = types.ContentTypeXML
+			contentType = types3.ContentTypeXML
 			reader = bytes.NewReader(content)
 			sha256Hex = utils.CalcSHA256Hex(content)
 		}
@@ -164,32 +205,32 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 
 	// set txn hash header
 	if txnHash != "" {
-		req.Header.Set(types.HTTPHeaderTransactionHash, txnHash)
+		req.Header.Set(types3.HTTPHeaderTransactionHash, txnHash)
 	}
 
 	// set content type header
 	if meta.contentType != "" {
-		req.Header.Set(types.HTTPHeaderContentType, meta.contentType)
+		req.Header.Set(types3.HTTPHeaderContentType, meta.contentType)
 	} else if contentType != "" {
-		req.Header.Set(types.HTTPHeaderContentType, contentType)
+		req.Header.Set(types3.HTTPHeaderContentType, contentType)
 	} else {
-		req.Header.Set(types.HTTPHeaderContentType, types.ContentDefault)
+		req.Header.Set(types3.HTTPHeaderContentType, types3.ContentDefault)
 	}
 
 	// set md5 header
 	if meta.contentMD5Base64 != "" {
-		req.Header[types.HTTPHeaderContentMD5] = []string{meta.contentMD5Base64}
+		req.Header[types3.HTTPHeaderContentMD5] = []string{meta.contentMD5Base64}
 	}
 
 	// set sha256 header
 	if meta.contentSHA256 != "" {
-		req.Header[types.HTTPHeaderContentSHA256] = []string{meta.contentSHA256}
+		req.Header[types3.HTTPHeaderContentSHA256] = []string{meta.contentSHA256}
 	} else {
-		req.Header[types.HTTPHeaderContentSHA256] = []string{sha256Hex}
+		req.Header[types3.HTTPHeaderContentSHA256] = []string{sha256Hex}
 	}
 
 	if meta.Range != "" && method == http.MethodGet {
-		req.Header.Set(types.HTTPHeaderRange, meta.Range)
+		req.Header.Set(types3.HTTPHeaderRange, meta.Range)
 	}
 
 	if isAdminAPi {
@@ -197,13 +238,13 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 		// if challengeInfo.ObjectId is not empty, other field should be set as well
 		if meta.challengeInfo.ObjectId != "" {
 			info := meta.challengeInfo
-			req.Header.Set(types.HTTPHeaderObjectId, info.ObjectId)
-			req.Header.Set(types.HTTPHeaderRedundancyIndex, strconv.Itoa(info.RedundancyIndex))
-			req.Header.Set(types.HTTPHeaderPieceIndex, strconv.Itoa(info.PieceIndex))
+			req.Header.Set(types3.HTTPHeaderObjectId, info.ObjectId)
+			req.Header.Set(types3.HTTPHeaderRedundancyIndex, strconv.Itoa(info.RedundancyIndex))
+			req.Header.Set(types3.HTTPHeaderPieceIndex, strconv.Itoa(info.PieceIndex))
 		}
 
 		if meta.TxnMsg != "" {
-			req.Header.Set(types.HTTPHeaderUnsignedMsg, meta.TxnMsg)
+			req.Header.Set(types3.HTTPHeaderUnsignedMsg, meta.TxnMsg)
 		}
 
 	} else {
@@ -217,18 +258,18 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 
 	if meta.userInfo.Address != "" {
 		info := meta.userInfo
-		req.Header.Set(types.HTTPHeaderUserAddress, info.Address)
+		req.Header.Set(types3.HTTPHeaderUserAddress, info.Address)
 	}
 
 	// set date header
 	stNow := time.Now().UTC()
-	req.Header.Set(types.HTTPHeaderDate, stNow.Format(types.Iso8601DateFormatSecond))
+	req.Header.Set(types3.HTTPHeaderDate, stNow.Format(types3.Iso8601DateFormatSecond))
 
 	// set user-agent
 	// req.Header.Set(types.HTTPHeaderUserAgent, c.userAgent)
 
 	// sign the total http request info when auth type v1
-	err = c.SignRequest(req, authInfo)
+	err = c.signRequest(req, authInfo)
 	if err != nil {
 		return req, err
 	}
@@ -237,7 +278,7 @@ func (c *Client) newRequest(ctx context.Context, method string, meta requestMeta
 }
 
 // doAPI call client.Do() to send request and read response from servers
-func (c *Client) doAPI(ctx context.Context, req *http.Request, meta requestMeta, closeBody bool) (*http.Response, error) {
+func (c *client) doAPI(ctx context.Context, req *http.Request, meta requestMeta, closeBody bool) (*http.Response, error) {
 	var cancel context.CancelFunc
 	if closeBody {
 		ctx, cancel = context.WithCancel(ctx)
@@ -281,7 +322,7 @@ func (c *Client) doAPI(ctx context.Context, req *http.Request, meta requestMeta,
 }
 
 // sendReq sends the message via REST and handles the response
-func (c *Client) sendReq(ctx context.Context, metadata requestMeta, opt *sendOptions, authInfo AuthInfo, endpoint *url.URL) (res *http.Response, err error) {
+func (c *client) sendReq(ctx context.Context, metadata requestMeta, opt *sendOptions, authInfo types3.AuthInfo, endpoint *url.URL) (res *http.Response, err error) {
 	req, err := c.newRequest(ctx, opt.method, metadata, opt.body, opt.txnHash, opt.isAdminApi, endpoint, authInfo)
 	if err != nil {
 		log.Debug().Msg("new request error stop send request" + err.Error())
@@ -296,8 +337,8 @@ func (c *Client) sendReq(ctx context.Context, metadata requestMeta, opt *sendOpt
 	return resp, nil
 }
 
-// GenerateURL constructs the target request url based on the parameters
-func (c *Client) GenerateURL(bucketName string, objectName string, relativePath string,
+// generateURL constructs the target request url based on the parameters
+func (c *client) generateURL(bucketName string, objectName string, relativePath string,
 	queryValues url.Values, isAdminApi bool, endpoint *url.URL) (*url.URL, error) {
 	host := endpoint.Host
 	scheme := endpoint.Scheme
@@ -314,7 +355,7 @@ func (c *Client) GenerateURL(bucketName string, objectName string, relativePath 
 
 	var urlStr string
 	if isAdminApi {
-		prefix := types.AdminURLPrefix + types.AdminURLVersion
+		prefix := types3.AdminURLPrefix + types3.AdminURLVersion
 		urlStr = scheme + "://" + host + prefix + "/"
 	} else {
 		// generate s3 virtual hosted style url, consider case where ListBuckets not having a bucket name
@@ -344,38 +385,35 @@ func (c *Client) GenerateURL(bucketName string, objectName string, relativePath 
 	return url.Parse(urlStr)
 }
 
-// SignRequest signs the request and set authorization before send to server
-func (c *Client) SignRequest(req *http.Request, info AuthInfo) error {
+// signRequest signs the request and set authorization before send to server
+func (c *client) signRequest(req *http.Request, info types3.AuthInfo) error {
 	var authStr []string
-	if info.SignType == types.AuthV1 {
+	if info.SignType == types3.AuthV1 {
 		signMsg := httplib.GetMsgToSign(req)
 
-		// TODO(leo) sign with new Account
-		/*
-			if c.signer == nil {
-				return errors.New("signer can not be nil with auth v1 type")
-			}
+		km, err := c.chainClient.GetKeyManager()
+		if err != nil {
+			return err
+		}
+		// sign the request header info, generate the signature
+		signature, err := km.Sign(signMsg)
+		if err != nil {
+			return err
+		}
 
-			// sign the request header info, generate the signature
-			signature, _, err := c.signer.Sign(signMsg)
-			if err != nil {
-				return err
-			}
-		*/
-		signature := []byte("")
 		authStr = []string{
-			types.AuthV1 + " " + types.SignAlgorithm,
+			types3.AuthV1 + " " + types3.SignAlgorithm,
 			" SignedMsg=" + hex.EncodeToString(signMsg),
 			"Signature=" + hex.EncodeToString(signature),
 		}
 
-	} else if info.SignType == types.AuthV2 {
+	} else if info.SignType == types3.AuthV2 {
 		if info.WalletSignStr == "" {
 			return errors.New("wallet signature can not be empty with auth v2 type")
 		}
 		// wallet should use same sign algorithm
 		authStr = []string{
-			types.AuthV2 + " " + types.SignAlgorithm,
+			types3.AuthV2 + " " + types3.SignAlgorithm,
 			" Signature=" + info.WalletSignStr,
 		}
 	} else {
@@ -383,14 +421,14 @@ func (c *Client) SignRequest(req *http.Request, info AuthInfo) error {
 	}
 
 	// set auth header
-	req.Header.Set(types.HTTPHeaderAuthorization, strings.Join(authStr, ", "))
+	req.Header.Set(types3.HTTPHeaderAuthorization, strings.Join(authStr, ", "))
 
 	return nil
 }
 
 // GetPieceHashRoots returns primary pieces, secondary piece Hash roots list and the object size
 // It is used for generate meta of object on the chain
-func (c *Client) GetPieceHashRoots(reader io.Reader, segSize int64,
+func (c *client) GetPieceHashRoots(reader io.Reader, segSize int64,
 	dataShards, parityShards int) ([]byte, [][]byte, int64, storageTypes.RedundancyType, error) {
 	pieceHashRoots, size, redundancyType, err := hashlib.ComputeIntegrityHash(reader, segSize, dataShards, parityShards)
 	if err != nil {
@@ -400,54 +438,32 @@ func (c *Client) GetPieceHashRoots(reader io.Reader, segSize int64,
 	return pieceHashRoots[0], pieceHashRoots[1:], size, redundancyType, nil
 }
 
-// NewAuthInfo returns the AuthInfo which need to pass to api
-// useWalletSign indicates whether you need use wallet to sign
-// signStr indicates the wallet signature or jwt token
-func NewAuthInfo(useWalletSign bool, signStr string) AuthInfo {
-	if !useWalletSign {
-		return AuthInfo{
-			SignType:      types.AuthV1,
-			WalletSignStr: "",
-		}
-	} else {
-		return AuthInfo{
-			SignType:      types.AuthV2,
-			WalletSignStr: signStr,
-		}
-	}
-}
-
-// NewStatement return the statement of permission module
-func (c *Client) NewStatement(actions []permTypes.ActionType, effect permTypes.Effect,
-	resource []string, opts client.NewStatementOptions) permTypes.Statement {
-	statement := permTypes.Statement{
-		Actions:        actions,
-		Effect:         effect,
-		Resources:      resource,
-		ExpirationTime: opts.StatementExpireTime,
+// sendPutPolicyTxn broadcast the putPolicy msg and return the txn hash
+func (c *client) sendPutPolicyTxn(msg *storageTypes.MsgPutPolicy, txOpts *types2.TxOption) (string, error) {
+	if err := msg.ValidateBasic(); err != nil {
+		return "", err
 	}
 
-	if opts.LimitSize != 0 {
-		statement.LimitSize = &common.UInt64Value{Value: opts.LimitSize}
-	}
-
-	return statement
-}
-
-func (c *Client) NewPrincipalWithAccount(principalAddr sdk.AccAddress) (client.Principal, error) {
-	p := permTypes.NewPrincipalWithAccount(principalAddr)
-	principalBytes, err := p.Marshal()
+	resp, err := c.chainClient.BroadcastTx([]sdk.Msg{msg}, txOpts)
 	if err != nil {
 		return "", err
 	}
-	return client.Principal(principalBytes), nil
+
+	return resp.TxResponse.TxHash, err
 }
 
-func (c *Client) NewPrincipalWithGroupId(groupId uint64) (client.Principal, error) {
-	p := permTypes.NewPrincipalWithGroup(sdkmath.NewUint(groupId))
-	principalBytes, err := p.Marshal()
+// sendDelPolicyTxn broadcast the deletePolicy msg and return the txn hash
+func (c *client) sendDelPolicyTxn(operator sdk.AccAddress, resource string, principal *permTypes.Principal, txOpts *types2.TxOption) (string, error) {
+	delPolicyMsg := storageTypes.NewMsgDeletePolicy(operator, resource, principal)
+
+	if err := delPolicyMsg.ValidateBasic(); err != nil {
+		return "", err
+	}
+
+	resp, err := c.chainClient.BroadcastTx([]sdk.Msg{delPolicyMsg}, txOpts)
 	if err != nil {
 		return "", err
 	}
-	return client.Principal(principalBytes), nil
+
+	return resp.TxResponse.TxHash, err
 }
