+++ client/api_object.go
@@ -1,13 +1,15 @@
-package api
+package client
 
 import (
 	"context"
+	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"math"
 	"net/http"
+	"net/url"
 	"os"
 	"strconv"
 	"strings"
@@ -15,20 +17,55 @@ import (
 	hashlib "github.com/bnb-chain/greenfield-common/go/hash"
 	sdkerror "github.com/bnb-chain/greenfield-go-sdk/pkg/error"
 	"github.com/bnb-chain/greenfield-go-sdk/pkg/utils"
+	"github.com/bnb-chain/greenfield-go-sdk/types"
+	gnfdSdkTypes "github.com/bnb-chain/greenfield/sdk/types"
 	gnfdTypes "github.com/bnb-chain/greenfield/types"
 	"github.com/bnb-chain/greenfield/types/s3util"
 	permTypes "github.com/bnb-chain/greenfield/x/permission/types"
 	storageTypes "github.com/bnb-chain/greenfield/x/storage/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/rs/zerolog/log"
-
-	"github.com/bnb-chain/greenfield-go-sdk/client"
-	"github.com/bnb-chain/greenfield-go-sdk/pkg/types"
 )
+
+type Object interface {
+	GetCreateObjectApproval(ctx context.Context, createObjectMsg *storageTypes.MsgCreateObject,
+		authInfo types.AuthInfo) (*storageTypes.MsgCreateObject, error)
+	CreateObject(ctx context.Context, bucketName, objectName string,
+		reader io.Reader, opts *CreateObjectOptions) (string, error)
+	PutObject(ctx context.Context, bucketName, objectName, txnHash string, objectSize int64,
+		reader io.Reader, opt types.PutObjectOption) (err error)
+	CancelCreateObject(bucketName, objectName string, opt types.CancelCreateOption) (string, error)
+	GetObject(ctx context.Context, bucketName, objectName string, opt types.GetObjectOption) (io.ReadCloser, GetObjectResult, error)
+	// HeadObject query the objectInfo on chain to check th object id, return the object info if exists
+	// return err info if object not exist
+	HeadObject(ctx context.Context, bucketName, objectName string) (*storageTypes.ObjectInfo, error)
+	// HeadObjectByID query the objectInfo on chain by object id, return the object info if exists
+	// return err info if object not exist
+	HeadObjectByID(ctx context.Context, objID string) (*storageTypes.ObjectInfo, error)
+	// PutObjectPolicy apply object policy to the principal, return the txn hash
+	PutObjectPolicy(bucketName, objectName string, principalStr types.Principal,
+		statements []*permTypes.Statement, opt types.PutPolicyOption) (string, error)
+	// GetObjectPolicy get the object policy info of the user specified by principalAddr
+	GetObjectPolicy(ctx context.Context, bucketName, objectName string, principalAddr sdk.AccAddress) (*permTypes.Policy, error)
+	ListObjects(ctx context.Context, bucketName string, authInfo types.AuthInfo) (ListObjectsResult, error)
+}
+
+// CreateObjectOptions indicates the metadata to construct `createObject` message of storage module
+type CreateObjectOptions struct {
+	Visibility      storageTypes.VisibilityType
+	TxOpts          *gnfdSdkTypes.TxOption
+	SecondarySPAccs []sdk.AccAddress
+	ContentType     string
+	IsReplicaType   bool // indicates whether the object use REDUNDANCY_REPLICA_TYPE
+}
+
+type DeleteObjectOption struct {
+	TxOpts *gnfdSdkTypes.TxOption
+}
 
 // GetRedundancyParams query and return the data shards, parity shards and segment size of redundancy
 // configuration on chain
-func (c *Client) GetRedundancyParams() (uint32, uint32, uint64, error) {
+func (c *client) GetRedundancyParams() (uint32, uint32, uint64, error) {
 	query := storageTypes.QueryParamsRequest{}
 	queryResp, err := c.chainClient.StorageQueryClient.Params(context.Background(), &query)
 	if err != nil {
@@ -40,7 +77,7 @@ func (c *Client) GetRedundancyParams() (uint32, uint32, uint64, error) {
 }
 
 // ComputeHashRoots return the hash roots list and content size
-func (c *Client) ComputeHashRoots(reader io.Reader) ([][]byte, int64, storageTypes.RedundancyType, error) {
+func (c *client) ComputeHashRoots(reader io.Reader) ([][]byte, int64, storageTypes.RedundancyType, error) {
 	dataBlocks, parityBlocks, segSize, err := c.GetRedundancyParams()
 	if err != nil {
 		return nil, 0, storageTypes.REDUNDANCY_EC_TYPE, err
@@ -51,8 +88,8 @@ func (c *Client) ComputeHashRoots(reader io.Reader) ([][]byte, int64, storageTyp
 }
 
 // CreateObject get approval of creating object and send createObject txn to greenfield chain
-func (c *Client) CreateObject(ctx context.Context, bucketName, objectName string,
-	reader io.Reader, opts client.CreateObjectOptions) (string, error) {
+func (c *client) CreateObject(ctx context.Context, bucketName, objectName string,
+	reader io.Reader, opts CreateObjectOptions) (string, error) {
 	if reader == nil {
 		return "", errors.New("fail to compute hash of payload, reader is nil")
 	}
@@ -96,7 +133,7 @@ func (c *Client) CreateObject(ctx context.Context, bucketName, objectName string
 		return "", err
 	}
 
-	signedCreateObjectMsg, err := c.GetCreateObjectApproval(ctx, createObjectMsg, NewAuthInfo(false, ""))
+	signedCreateObjectMsg, err := c.GetCreateObjectApproval(ctx, createObjectMsg, types.NewAuthInfo(false, ""))
 	if err != nil {
 		return "", err
 	}
@@ -109,7 +146,7 @@ func (c *Client) CreateObject(ctx context.Context, bucketName, objectName string
 }
 
 // DeleteObject send DeleteBucket txn to greenfield chain and return txn hash
-func (c *Client) DeleteObject(bucketName, objectName string, opt client.DeleteObjectOption) (string, error) {
+func (c *client) DeleteObject(bucketName, objectName string, opt DeleteObjectOption) (string, error) {
 	if err := s3util.CheckValidBucketName(bucketName); err != nil {
 		return "", err
 	}
@@ -133,7 +170,7 @@ func (c *Client) DeleteObject(bucketName, objectName string, opt client.DeleteOb
 }
 
 // CancelCreateObject send CancelCreateObject txn to greenfield chain
-func (c *Client) CancelCreateObject(bucketName, objectName string, opt client.CancelCreateOption) (string, error) {
+func (c *client) CancelCreateObject(bucketName, objectName string, opt types.CancelCreateOption) (string, error) {
 	if err := s3util.CheckValidBucketName(bucketName); err != nil {
 		return "", err
 	}
@@ -159,8 +196,8 @@ func (c *Client) CancelCreateObject(bucketName, objectName string, opt client.Ca
 
 // PutObject supports the second stage of uploading the object to bucket.
 // txnHash should be the str which hex.encoding from txn hash bytes
-func (c *Client) PutObject(ctx context.Context, bucketName, objectName, txnHash string, objectSize int64,
-	reader io.Reader, authInfo AuthInfo, opt client.PutObjectOption,
+func (c *client) PutObject(ctx context.Context, bucketName, objectName, txnHash string, objectSize int64,
+	reader io.Reader, authInfo types.AuthInfo, opt types.PutObjectOption,
 ) (err error) {
 	if txnHash == "" {
 		return errors.New("txn hash empty")
@@ -205,8 +242,8 @@ func (c *Client) PutObject(ctx context.Context, bucketName, objectName, txnHash 
 }
 
 // FPutObject supports uploading object from local file
-func (c *Client) FPutObject(ctx context.Context, bucketName, objectName,
-	filePath, txnHash, contentType string, authInfo AuthInfo,
+func (c *client) FPutObject(ctx context.Context, bucketName, objectName,
+	filePath, txnHash, contentType string, authInfo types.AuthInfo,
 ) (err error) {
 	fReader, err := os.Open(filePath)
 	// If any error fail quickly here.
@@ -221,18 +258,18 @@ func (c *Client) FPutObject(ctx context.Context, bucketName, objectName,
 		return err
 	}
 
-	return c.PutObject(ctx, bucketName, objectName, txnHash, stat.Size(), fReader, authInfo, client.PutObjectOption{ContentType: contentType})
+	return c.PutObject(ctx, bucketName, objectName, txnHash, stat.Size(), fReader, authInfo, types.PutObjectOption{ContentType: contentType})
 }
 
 // GetObject download s3 object payload and return the related object info
-func (c *Client) GetObject(ctx context.Context, bucketName, objectName string,
-	opts client.GetObjectOption, authInfo AuthInfo) (io.ReadCloser, ObjectInfo, error) {
+func (c *client) GetObject(ctx context.Context, bucketName, objectName string,
+	opts types.GetObjectOption, authInfo types.AuthInfo) (io.ReadCloser, types.ObjectStat, error) {
 	if err := s3util.CheckValidBucketName(bucketName); err != nil {
-		return nil, ObjectInfo{}, err
+		return nil, types.ObjectStat{}, err
 	}
 
 	if err := s3util.CheckValidObjectName(objectName); err != nil {
-		return nil, ObjectInfo{}, err
+		return nil, types.ObjectStat{}, err
 	}
 
 	reqMeta := requestMeta{
@@ -252,25 +289,25 @@ func (c *Client) GetObject(ctx context.Context, bucketName, objectName string,
 
 	endpoint, err := c.getSPUrlByBucket(bucketName)
 	if err != nil {
-		return nil, ObjectInfo{}, err
+		return nil, types.ObjectStat{}, err
 	}
 
 	resp, err := c.sendReq(ctx, reqMeta, &sendOpt, authInfo, endpoint)
 	if err != nil {
-		return nil, ObjectInfo{}, err
+		return nil, types.ObjectStat{}, err
 	}
 
-	ObjInfo, err := getObjInfo(objectName, resp.Header)
+	objStat, err := getObjInfo(objectName, resp.Header)
 	if err != nil {
 		utils.CloseResponse(resp)
-		return nil, ObjectInfo{}, err
+		return nil, types.ObjectStat{}, err
 	}
 
-	return resp.Body, ObjInfo, nil
+	return resp.Body, objStat, nil
 }
 
 // FGetObject download s3 object payload adn write the object content into local file specified by filePath
-func (c *Client) FGetObject(ctx context.Context, bucketName, objectName, filePath string, opts client.GetObjectOption, authinfo AuthInfo) error {
+func (c *client) FGetObject(ctx context.Context, bucketName, objectName, filePath string, opts types.GetObjectOption, authInfo types.AuthInfo) error {
 	// Verify if destination already exists.
 	st, err := os.Stat(filePath)
 	if err == nil {
@@ -286,7 +323,7 @@ func (c *Client) FGetObject(ctx context.Context, bucketName, objectName, filePat
 		return err
 	}
 
-	body, _, err := c.GetObject(ctx, bucketName, objectName, opts, authinfo)
+	body, _, err := c.GetObject(ctx, bucketName, objectName, opts, authInfo)
 	if err != nil {
 		return err
 	}
@@ -302,7 +339,7 @@ func (c *Client) FGetObject(ctx context.Context, bucketName, objectName, filePat
 }
 
 // getObjInfo generates objectInfo base on the response http header content
-func getObjInfo(objectName string, h http.Header) (ObjectInfo, error) {
+func getObjInfo(objectName string, h http.Header) (types.ObjectStat, error) {
 	// Parse content length is exists
 	var size int64 = -1
 	var err error
@@ -310,7 +347,7 @@ func getObjInfo(objectName string, h http.Header) (ObjectInfo, error) {
 	if contentLength != "" {
 		size, err = strconv.ParseInt(contentLength, 10, 64)
 		if err != nil {
-			return ObjectInfo{}, sdkerror.ErrResponse{
+			return types.ObjectStat{}, sdkerror.ErrResponse{
 				Code:    "InternalError",
 				Message: fmt.Sprintf("Content-Length parse error %v", err),
 			}
@@ -323,7 +360,7 @@ func getObjInfo(objectName string, h http.Header) (ObjectInfo, error) {
 		contentType = types.ContentDefault
 	}
 
-	return ObjectInfo{
+	return types.ObjectStat{
 		ObjectName:  objectName,
 		ContentType: contentType,
 		Size:        size,
@@ -332,7 +369,7 @@ func getObjInfo(objectName string, h http.Header) (ObjectInfo, error) {
 
 // HeadObject query the objectInfo on chain to check th object id, return the object info if exists
 // return err info if object not exist
-func (c *Client) HeadObject(ctx context.Context, bucketName, objectName string) (*storageTypes.ObjectInfo, error) {
+func (c *client) HeadObject(ctx context.Context, bucketName, objectName string) (*storageTypes.ObjectInfo, error) {
 	queryHeadObjectRequest := storageTypes.QueryHeadObjectRequest{
 		BucketName: bucketName,
 		ObjectName: objectName,
@@ -347,7 +384,7 @@ func (c *Client) HeadObject(ctx context.Context, bucketName, objectName string) 
 
 // HeadObjectByID query the objectInfo on chain by object id, return the object info if exists
 // return err info if object not exist
-func (c *Client) HeadObjectByID(ctx context.Context, objID string) (*storageTypes.ObjectInfo, error) {
+func (c *client) HeadObjectByID(ctx context.Context, objID string) (*storageTypes.ObjectInfo, error) {
 	headObjectRequest := storageTypes.QueryHeadObjectByIdRequest{
 		ObjectId: objID,
 	}
@@ -360,8 +397,8 @@ func (c *Client) HeadObjectByID(ctx context.Context, objID string) (*storageType
 }
 
 // PutObjectPolicy apply object policy to the principal, return the txn hash
-func (c *Client) PutObjectPolicy(bucketName, objectName string, principalStr client.Principal,
-	statements []*permTypes.Statement, opt client.PutPolicyOption) (string, error) {
+func (c *client) PutObjectPolicy(bucketName, objectName string, principalStr types.Principal,
+	statements []*permTypes.Statement, opt types.PutPolicyOption) (string, error) {
 	km, err := c.chainClient.GetKeyManager()
 	if err != nil {
 		return "", errors.New("key manager is nil")
@@ -380,7 +417,7 @@ func (c *Client) PutObjectPolicy(bucketName, objectName string, principalStr cli
 	return c.sendPutPolicyTxn(putPolicyMsg, opt.TxOpts)
 }
 
-func (c *Client) DeleteObjectPolicy(bucketName, objectName string, principalAddr sdk.AccAddress, opt client.DeletePolicyOption) (string, error) {
+func (c *client) DeleteObjectPolicy(bucketName, objectName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error) {
 	km, err := c.chainClient.GetKeyManager()
 	if err != nil {
 		return "", errors.New("key manager is nil")
@@ -392,7 +429,7 @@ func (c *Client) DeleteObjectPolicy(bucketName, objectName string, principalAddr
 }
 
 // IsObjectPermissionAllowed check if the permission of the object is allowed to the user
-func (c *Client) IsObjectPermissionAllowed(ctx context.Context, user sdk.AccAddress,
+func (c *client) IsObjectPermissionAllowed(ctx context.Context, user sdk.AccAddress,
 	bucketName, objectName string, action permTypes.ActionType) (permTypes.Effect, error) {
 	verifyReq := storageTypes.QueryVerifyPermissionRequest{
 		Operator:   user.String(),
@@ -410,7 +447,7 @@ func (c *Client) IsObjectPermissionAllowed(ctx context.Context, user sdk.AccAddr
 }
 
 // GetObjectPolicy get the object policy info of the user specified by principalAddr
-func (c *Client) GetObjectPolicy(ctx context.Context, bucketName, objectName string, principalAddr sdk.AccAddress) (*permTypes.Policy, error) {
+func (c *client) GetObjectPolicy(ctx context.Context, bucketName, objectName string, principalAddr sdk.AccAddress) (*permTypes.Policy, error) {
 	resource := gnfdTypes.NewObjectGRN(bucketName, objectName)
 	queryPolicy := storageTypes.QueryPolicyForAccountRequest{
 		Resource:         resource.String(),
@@ -426,9 +463,9 @@ func (c *Client) GetObjectPolicy(ctx context.Context, bucketName, objectName str
 }
 
 // ListObjects return object list of the specific bucket
-func (c *Client) ListObjects(ctx context.Context, bucketName string, authInfo AuthInfo) (client.ListObjectsResponse, error) {
+func (c *client) ListObjects(ctx context.Context, bucketName string, authInfo types.AuthInfo) (ListObjectsResult, error) {
 	if err := s3util.CheckValidBucketName(bucketName); err != nil {
-		return client.ListObjectsResponse{}, err
+		return ListObjectsResult{}, err
 	}
 
 	reqMeta := requestMeta{
@@ -443,31 +480,78 @@ func (c *Client) ListObjects(ctx context.Context, bucketName string, authInfo Au
 
 	endpoint, err := c.getSPUrlByBucket(bucketName)
 	if err != nil {
-		return client.ListObjectsResponse{}, err
+		return ListObjectsResult{}, err
 	}
 
 	resp, err := c.sendReq(ctx, reqMeta, &sendOpt, authInfo, endpoint)
 	if err != nil {
-		return client.ListObjectsResponse{}, err
+		return ListObjectsResult{}, err
 	}
 	defer utils.CloseResponse(resp)
 
-	ListObjectsResult := client.ListObjectsResponse{}
 	// unmarshal the json content from response body
 	buf := new(strings.Builder)
 	_, err = io.Copy(buf, resp.Body)
 	if err != nil {
 		log.Error().Msg("the list of objects in user's bucket:" + bucketName + " failed: " + err.Error())
-		return client.ListObjectsResponse{}, err
+		return ListObjectsResult{}, err
 	}
 
+	listObjectsResult := ListObjectsResult{}
 	bufStr := buf.String()
-	err = json.Unmarshal([]byte(bufStr), &ListObjectsResult)
+	err = json.Unmarshal([]byte(bufStr), &listObjectsResult)
 	//TODO(annie) remove tolerance for unmarshal err after structs got stabilized
-	if err != nil && ListObjectsResult.Objects == nil {
+	if err != nil && listObjectsResult.Objects == nil {
 		log.Error().Msg("the list of objects in user's bucket:" + bucketName + " failed: " + err.Error())
-		return client.ListObjectsResponse{}, err
+		return ListObjectsResult{}, err
 	}
 
-	return ListObjectsResult, nil
+	return listObjectsResult, nil
+}
+
+// GetCreateObjectApproval returns the signature info for the approval of preCreating resources
+func (c *client) GetCreateObjectApproval(ctx context.Context, createObjectMsg *storageTypes.MsgCreateObject, authInfo types.AuthInfo) (*storageTypes.MsgCreateObject, error) {
+	unsignedBytes := createObjectMsg.GetSignBytes()
+
+	// set the action type
+	urlVal := make(url.Values)
+	urlVal["action"] = []string{types.CreateObjectAction}
+
+	reqMeta := requestMeta{
+		urlValues:     urlVal,
+		urlRelPath:    "get-approval",
+		contentSHA256: types.EmptyStringSHA256,
+		TxnMsg:        hex.EncodeToString(unsignedBytes),
+	}
+
+	sendOpt := sendOptions{
+		method:     http.MethodGet,
+		isAdminApi: true,
+	}
+
+	endpoint, err := c.getSPUrlByBucket(createObjectMsg.BucketName)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := c.sendReq(ctx, reqMeta, &sendOpt, authInfo, endpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	// fetch primary signed msg from sp response
+	signedRawMsg := resp.Header.Get(types.HTTPHeaderSignedMsg)
+	if signedRawMsg == "" {
+		return nil, errors.New("fail to fetch pre createObject signature")
+	}
+
+	signedMsgBytes, err := hex.DecodeString(signedRawMsg)
+	if err != nil {
+		return nil, err
+	}
+
+	var signedMsg storageTypes.MsgCreateObject
+	storageTypes.ModuleCdc.MustUnmarshalJSON(signedMsgBytes, &signedMsg)
+
+	return &signedMsg, nil
 }
