--- pkg/api/api_bank.go
@@ -1,5 +1,0 @@
-package api
-
-func (c Client) Send()          {}
-func (c Client) MultiSend()     {}
-func (c Client) QueryBalances() {}
